Make it work with TensorFlow r1.9
* `inter_op_parallelism_threads` and `intra_op_parallelism_threads`
options are added to `neural_network`. This options is applied to
`tf.ConfigProto`. When they are set to zero, the number of threads
should be decided automatically. However, since it is not working on
TensorFlow r1.9 and defaults to 2, we added options to set these values.
Default is set to zero.

* `tf.losses.get_regularization_loss()` is returning `float32` as of
TensorFlow r1.9. To get around this, we cast this value to `float64`.